move accentuate test out of spec into dedicated shared file
